🚀 Add Makefile and Refactor Audio Components for Resonate Protocol 🚀
- **Created a Makefile** to automate build processes for the Resonate Protocol server and player with targets for building, testing, cleaning, and installation.
- **Refactored log file handling** in `main.go` to use `os.O_WRONLY`, ensuring that the logging file isn't unnecessarily read for output.
- **Altered PCM sample handling** in `decoder.go` to use `[]int16` instead of `[]byte`, improving performance by avoiding conversions.
- **Updated audio output** in `output.go` to utilize a pipe for continuous streaming instead of repeatedly creating players for each buffer, enhancing efficiency during playback.
- **Implemented buffering and playback initialization** in `scheduler.go` to ensure a smoother audio streaming experience with a buffering strategy that preloads before playback starts.
- **Refined audio encoding processes** in `audio_engine.go` to efficiently handle codecs and improve the clarity of PCM encoding.
- **Enhanced client handling in `server.go`** to manage duplicate connections gracefully, allowing the system to handle multiple clients without crashing.

🌟 These enhancements aim to improve the efficiency, maintainability, and user experience of the Resonate audio streaming application. Check out the `Makefile` for easy builds! 💪

+++ internal/audio/types.go
@@ -17,6 +17,6 @@ type Format struct {
 type Buffer struct {
 	Timestamp  int64     // Server timestamp (microseconds)
 	PlayAt     time.Time // Local play time
-	Samples    []byte    // PCM samples
+	Samples    []int16   // PCM samples (int16 format to avoid conversions)
 	Format     Format
 }
